:sparkles: feat: ColorModalForm 생성
ColorModal 에서 form 항목을 분리하여 컴포넌트로 생성. 온보딩 모달 제작에 사용하기 위함이며, 온보달 옵션 활성화일때만
tranlate 스타일속성을 부여
